+++ hydra-node/src/Hydra/Ledger/Cardano/Builder.hs
@@ -91,9 +91,9 @@ addOutputs :: [TxOut CtxTx] -> TxBuilder -> TxBuilder
 addOutputs outputs tx =
   tx{txOuts = txOuts tx <> outputs}
 
--- | Mint tokens identified by the given Plutus script.
-mintTokens :: PlutusScript -> AssetName -> Quantity -> TxBuilder -> TxBuilder
-mintTokens script assetName quantity tx =
+-- | Mint tokens with given plutus minting script and redeemer.
+mintTokens :: ToScriptData redeemer => PlutusScript -> redeemer -> [(AssetName, Quantity)] -> TxBuilder -> TxBuilder
+mintTokens script redeemer assets tx =
   tx{txMintValue = TxMintValue mintedTokens' mintedWitnesses'}
  where
   (mintedTokens, mintedWitnesses) =
@@ -104,19 +104,19 @@ mintTokens script assetName quantity tx =
         (t, m)
 
   mintedTokens' =
-    mintedTokens <> valueFromList [(AssetId policyId assetName, quantity)]
+    mintedTokens <> valueFromList (fmap (first (AssetId policyId)) assets)
 
   mintedWitnesses' =
     BuildTxWith $ mintedWitnesses <> Map.singleton policyId mintingWitness
 
   mintingWitness =
-    mkScriptWitness script NoScriptDatumForMint (toScriptData ())
+    mkScriptWitness script NoScriptDatumForMint (toScriptData redeemer)
 
   policyId =
     PolicyId $ hashScript $ PlutusScript script
 
--- | Burn tokens for given policy.
+-- | Burn tokens with given plutus minting script and redeemer.
 -- This is really just `mintTokens` with negated 'Quantity'.
-burnTokens :: PlutusScript -> AssetName -> Quantity -> TxBuilder -> TxBuilder
-burnTokens script assetName quantity =
-  mintTokens script assetName (negate quantity)
+burnTokens :: ToScriptData redeemer => PlutusScript -> redeemer -> [(AssetName, Quantity)] -> TxBuilder -> TxBuilder
+burnTokens script redeemer assets =
+  mintTokens script redeemer (fmap (second negate) assets)