+++ hydra-node/src/Hydra/Ledger/Cardano/Evaluate.hs
@@ -1,3 +1,11 @@
+-- | Simplified interface to phase-2 validation of transaction, eg. evaluation of Plutus scripts.
+--
+-- The `evaluateTx` function simplifies the call to underlying Plutus providing execution report
+-- using pre-canned `PParams`. This should only be used for /testing/ or /benchmarking/ purpose
+-- as the real evaluation parameters are set when the Hydra node starts.
+--
+-- __NOTE__: The reason this module is here instead of part of `test/` directory is to be used
+-- in @tx-cost@ executable.
 module Hydra.Ledger.Cardano.Evaluate where
 
 import Hydra.Prelude hiding (label)
@@ -51,8 +59,8 @@ pparams =
   def
     { _costmdls = Map.singleton PlutusV1 $ fromJust $ defaultCostModel PlutusV1
     , _maxValSize = 1000000000
-    , _maxTxExUnits = ExUnits 12_500_000 10_000_000_000
-    , _maxBlockExUnits = ExUnits 10000000000 10000000000
+    , _maxTxExUnits = ExUnits 14_000_000 10_000_000_000
+    , _maxBlockExUnits = ExUnits 56_000_000 40_000_000_000
     , _protocolVersion = ProtVer 5 0
     , _prices =
         Prices