+++ hydra-node/test/Hydra/Chain/Direct/Contract/Abort.hs
@@ -11,18 +11,25 @@ import qualified Cardano.Api.UTxO as UTxO
 import qualified Data.Map as Map
 import Data.Maybe (fromJust)
 import Hydra.Chain (HeadParameters (..))
-import Hydra.Chain.Direct.Contract.Mutation (Mutation (ChangeHeadDatum, ChangeInput, RemoveOutput), SomeMutation (..), anyPayToPubKeyTxOut, headTxIn)
-import Hydra.Chain.Direct.Fixture (testNetworkId, testPolicyId)
+import Hydra.Chain.Direct.Contract.Mutation (
+  Mutation (ChangeHeadDatum, ChangeInput, RemoveOutput),
+  SomeMutation (..),
+  addPTWithQuantity,
+  anyPayToPubKeyTxOut,
+  changeMintedValueQuantityFrom,
+  headTxIn,
+ )
+import Hydra.Chain.Direct.Fixture (testNetworkId, testPolicyId, testSeedInput)
 import qualified Hydra.Chain.Direct.Fixture as Fixture
-import Hydra.Chain.Direct.Tx (UTxOWithScript, abortTx, mkHeadOutputInitial)
-import Hydra.Chain.Direct.TxSpec (drop2nd, drop3rd, genAbortableOutputs)
+import Hydra.Chain.Direct.Tx (UTxOWithScript, abortTx, mkHeadOutputInitial, mkHeadTokenScript)
+import Hydra.Chain.Direct.TxSpec (drop3rd, genAbortableOutputs)
 import qualified Hydra.Contract.Commit as Commit
 import qualified Hydra.Contract.HeadState as Head
 import qualified Hydra.Contract.Initial as Initial
 import Hydra.Data.Party (partyFromVerKey)
 import Hydra.Party (Party, vkey)
 import Hydra.Prelude
-import Test.QuickCheck (Property, choose, counterexample, oneof)
+import Test.QuickCheck (Property, choose, counterexample, oneof, suchThat)
 
 --
 -- AbortTx
@@ -41,10 +48,13 @@ healthyAbortTx =
     abortTx
       Fixture.testNetworkId
       (headInput, toUTxOContext headOutput, headDatum)
-      (Map.fromList (drop2nd <$> healthyInitials))
+      headTokenScript
+      (Map.fromList (tripleToPair <$> healthyInitials))
       (Map.fromList (tripleToPair <$> healthyCommits))
 
   headInput = generateWith arbitrary 42
+
+  headTokenScript = mkHeadTokenScript testSeedInput
 
   headOutput = mkHeadOutputInitial testNetworkId testPolicyId headParameters
 
@@ -65,11 +75,13 @@ healthyAbortTx =
 healthyInitials :: [UTxOWithScript]
 healthyCommits :: [UTxOWithScript]
 (healthyInitials, healthyCommits) =
-  -- NOTE: Why 43 one may ask? Because 42 does not generate commit UTXOs
   -- TODO: Refactor this to be an AbortTx generator because we actually want
   -- to test healthy abort txs with varied combinations of inital and commit
   -- outputs
-  generateWith (genAbortableOutputs healthyParties) 43
+  generateWith (genAbortableOutputs healthyParties `suchThat` thereIsAtLeastOneCommit) 42
+
+thereIsAtLeastOneCommit :: ([UTxOWithScript], [UTxOWithScript]) -> Bool
+thereIsAtLeastOneCommit (is, cs) = not (null cs) && not (null is)
 
 healthyParties :: [Party]
 healthyParties =
@@ -98,20 +110,23 @@ propHasCommit (_, utxo) =
 
 data AbortMutation
   = MutateParties
-  | MutateDropCommitOutput
+  | DropOneCommitOutput
   | MutateHeadScriptInput
+  | BurnOneTokenMore
+  | MutateThreadTokenQuantity
   deriving (Generic, Show, Enum, Bounded)
 
 genAbortMutation :: (Tx, UTxO) -> Gen SomeMutation
-genAbortMutation (_, utxo) =
+genAbortMutation (tx, utxo) =
   oneof
     [ SomeMutation MutateParties . ChangeHeadDatum <$> do
         moreParties <- (: healthyParties) <$> arbitrary
         c <- arbitrary
         pure $ Head.Initial c (partyFromVerKey . vkey <$> moreParties)
-    , SomeMutation MutateDropCommitOutput
+    , SomeMutation DropOneCommitOutput
         . RemoveOutput
-        . (+ 1) -- NOTE(AB): Assumes the transaction's first output is the head output
-        <$> choose (0, fromIntegral (length healthyCommits - 1))
+        <$> choose (0, fromIntegral (length (txOuts' tx) - 1))
     , SomeMutation MutateHeadScriptInput . ChangeInput (headTxIn utxo) <$> anyPayToPubKeyTxOut
+    , SomeMutation MutateThreadTokenQuantity <$> changeMintedValueQuantityFrom tx (-1)
+    , SomeMutation BurnOneTokenMore <$> addPTWithQuantity tx (-1)
     ]