+++ hydra-node/test/Hydra/Chain/Direct/Fixture.hs
@@ -3,16 +3,17 @@
 
 -- | Unit tests for our "hand-rolled" transactions as they are used in the
 -- "direct" chain component.
-module Hydra.Chain.Direct.Fixture where
+module Hydra.Chain.Direct.Fixture (
+  module Hydra.Chain.Direct.Fixture,
+  pparams,
+) where
 
 import Hydra.Prelude
 
 import Cardano.Ledger.Alonzo.Language (Language (PlutusV1))
-import Cardano.Ledger.Alonzo.PParams (PParams, PParams' (..))
 import qualified Cardano.Ledger.Alonzo.PParams as Ledger.Alonzo
-import Cardano.Ledger.Alonzo.Scripts (CostModel, ExUnits (..), Prices (..))
+import Cardano.Ledger.Alonzo.Scripts (CostModel)
 import qualified Cardano.Ledger.Alonzo.Scripts as Ledger.Alonzo
-import Cardano.Ledger.BaseTypes (ProtVer (..), boundRational)
 import qualified Cardano.Ledger.BaseTypes as Ledger
 import qualified Cardano.Ledger.Shelley.Rules.Ledger as Ledger
 import qualified Cardano.Ledger.Slot as Ledger
@@ -26,7 +27,6 @@ import Data.Bits (shift)
 import Data.Default (def)
 import qualified Data.Map as Map
 import Data.Maybe (fromJust)
-import Data.Ratio ((%))
 import Data.Time.Clock.POSIX (posixSecondsToUTCTime)
 import Hydra.Cardano.Api (
   LedgerEra,
@@ -37,7 +37,7 @@ import Hydra.Cardano.Api (
   TxIn,
  )
 import Hydra.Chain.Direct.Tx (headPolicyId)
-import Hydra.Chain.Direct.Util (Era)
+import Hydra.Ledger.Cardano.Evaluate (pparams)
 import Plutus.V1.Ledger.Api (PubKeyHash (PubKeyHash), toBuiltin)
 import qualified Test.Cardano.Ledger.Alonzo.AlonzoEraGen as Ledger.Alonzo
 import Test.Cardano.Ledger.Alonzo.PlutusScripts (defaultCostModel)
@@ -55,21 +55,6 @@ testSeedInput = generateWith arbitrary 42
 
 maxTxSize :: Int64
 maxTxSize = 1 `shift` 15
-
-pparams :: PParams Era
-pparams =
-  def
-    { _costmdls = Map.singleton PlutusV1 $ fromJust defaultCostModel
-    , _maxValSize = 1000000000
-    , _maxTxExUnits = ExUnits 12_500_000 10_000_000_000
-    , _maxBlockExUnits = ExUnits 10000000000 10000000000
-    , _protocolVersion = ProtVer 5 0
-    , _prices =
-        Prices
-          { prMem = fromJust $ boundRational $ 721 % 10000000
-          , prSteps = fromJust $ boundRational $ 577 % 10000
-          }
-    }
 
 instance Arbitrary PubKeyHash where
   arbitrary = PubKeyHash . toBuiltin <$> (arbitrary :: Gen ByteString)