+++ CHANGELOG.md
@@ -13,6 +13,7 @@ changes.
 #### Added
 
 - Support multiple Heads per Cardano network by identifying and distinguishing transactions of individual Head instances [#180](https://github.com/input-output-hk/hydra-poc/issues/180).
+- Mint and burn state token used to thread state across the OCV state machine, and participation tokens for each party in the head [#181](https://github.com/input-output-hk/hydra-poc/issues/181).
 
 #### Changed